pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodub-pro"
version = "0.1.0"
description = "Automated video dubbing solution with AI"
readme = "README.md"
authors = [
    {name = "AutoDub Team", email = "example@autodub.pro"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.37.37",
    "python-dotenv>=1.0.0",
    "joblib>=1.4.2",
    "moviepy>=2.1.2",
    "mutagen>=1.47.0",
    "numpy>=2.2.5",
    "openai>=1.75.0",
    "pydrive>=1.3.1",
    "pydub>=0.25.1",
    "pyside6>=6.9.0",
    "pysrt>=1.1.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "soundfile>=0.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/autodub-pro"
"Bug Tracker" = "https://github.com/yourusername/autodub-pro/issues"

[project.scripts]
autodub = "autodub_pro.main:main"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
